Reorganised file and folder structure for LandingPage, plus added Program component
