+++ client/life-goats-app/src/components/LandingPage/parallax/Parallax.jsx
