+++ client/life-goats-app/src/components/LandingPage/program/Program.jsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const Program = () => {
+  return (
+    <div>
+      <h1>Program</h1>
+    </div>
+  )
+}
+
+export default Program;
